ui: Re-adjust ui elements in overviewCard
