+++ src/components/stickable/list/CardDecoration.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import { StickableStatus } from '@/lib/utils';
-import BadgeRow from './BadgeRow.vue';
 import FavoriteMark from './FavoriteMark.vue';
+import StatusIcon from './StatusIcon.vue';
 
 const props = defineProps<{
   isFavorite: boolean;
@@ -18,6 +18,6 @@ const props = defineProps<{
     :status="props.status"
   >
     <FavoriteMark :show="props.isFavorite" />
-    <BadgeRow :isNew="props.isNew" :status="props.status" />
+    <StatusIcon :isNew="props.isNew" :status="props.status" />
   </slot>
 </template>
