+++ src/components/stickable/list/FavoriteMark.vue
@@ -6,6 +6,6 @@ defineProps<{ show: boolean }>();
 
 <template>
   <div v-if="show">
-    <Icon icon="ic:round-star" class="absolute top-1 right-1 h-5 w-5 z-10" />
+    <Icon icon="ic:round-star" class="absolute top-1 left-1 h-5 w-5 z-10" />
   </div>
 </template>
