+++ src/routes/tricks/TrickList.vue
@@ -170,7 +170,7 @@ function linkToDetails(primaryKey: PrimaryKey): string {
         >
           <div class="text-lg font-medium px-3 w-full text-center">{{ section.title }}</div>
           <div
-            class="grid grid-cols-3 md:grid-cols-4 lg:grid-cols-5 xl:grid-cols-6 gap-2 w-full grid-flow-row-dense"
+            class="grid grid-cols-3 sm:grid-cols-4 md:grid-cols-5 lg:grid-cols-6 xl:grid-cols-8 gap-2 w-full grid-flow-row-dense"
           >
             <StickableOverviewCard
               v-for="item in section.items"
